requirements.txt
github3.py==0.6.1
requests==1.1.0
wsgiref==0.1.2
